refactor: repository 래픽토링 및 도메인 분리
