--- clokey-api/src/main/java/org/clokey/domain/cloth/repository/ClothImageRepository.java
@@ -1,7 +1,0 @@
-package org.clokey.domain.cloth.repository;
-
-import org.clokey.cloth.entity.ClothImage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ClothImageRepository
-        extends JpaRepository<ClothImage, Long>, ClothProjectionRepository {}
