--- clokey-api/src/main/java/org/clokey/domain/cloth/repository/ClothProjectionRepository.java
@@ -1,3 +1,0 @@
-package org.clokey.domain.cloth.repository;
-
-public interface ClothProjectionRepository {}
