+++ clokey-api/src/main/java/org/clokey/domain/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package org.clokey.domain.history.repository;
+package org.clokey.domain.comment.repository;
 
-import org.clokey.history.entity.Comment;
+import org.clokey.comment.entitiy.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {}
