--- clokey-api/src/main/java/org/clokey/domain/history/repository/HistoryProjectionRepository.java
@@ -1,3 +1,0 @@
-package org.clokey.domain.history.repository;
-
-public interface HistoryProjectionRepository {}
