+++ clokey-api/src/main/java/org/clokey/domain/history/repository/HistoryRepository.java
@@ -3,5 +3,4 @@ package org.clokey.domain.history.repository;
 import org.clokey.history.entity.History;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoryRepository
-        extends JpaRepository<History, Long>, HistoryProjectionRepository {}
+public interface HistoryRepository extends JpaRepository<History, Long> {}
