--- clokey-api/src/main/java/org/clokey/domain/member/repository/MemberProjectionRepository.java
@@ -1,3 +1,0 @@
-package org.clokey.domain.member.repository;
-
-public interface MemberProjectionRepository {}
