+++ clokey-api/src/main/java/org/clokey/domain/member/repository/MemberRepository.java
@@ -3,4 +3,4 @@ package org.clokey.domain.member.repository;
 import org.clokey.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberProjectionRepository {}
+public interface MemberRepository extends JpaRepository<Member, Long> {}
