+++ clokey-api/src/main/java/org/clokey/domain/notification/repository/ClokeyNotificationRepository.java
@@ -3,4 +3,4 @@ package org.clokey.domain.notification.repository;
 import org.clokey.notification.entity.ClokeyNotification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<ClokeyNotification, Long> {}
+public interface ClokeyNotificationRepository extends JpaRepository<ClokeyNotification, Long> {}
