--- clokey-api/src/main/java/org/clokey/domain/recommendation/repository/RecommendationRepository.java
@@ -1,6 +1,0 @@
-package org.clokey.domain.recommendation.repository;
-
-import org.clokey.recommendation.entity.Recommendation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecommendationRepository extends JpaRepository<Recommendation, Long> {}
