+++ clokey-api/src/main/java/org/clokey/domain/report/repository/CommentReportRepository.java
@@ -2,7 +2,5 @@ package org.clokey.domain.report.repository;
 
 import org.clokey.report.entity.CommentReport;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface CommentReportRepository
-        extends JpaRepository<CommentReport, Long>, QuerydslPredicateExecutor<CommentReport> {}
+public interface CommentReportRepository extends JpaRepository<CommentReport, Long> {}
