+++ clokey-api/src/main/java/org/clokey/domain/report/repository/ProfileReportRepository.java
@@ -2,7 +2,5 @@ package org.clokey.domain.report.repository;
 
 import org.clokey.report.entity.ProfileReport;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface ProfileReportRepository
-        extends JpaRepository<ProfileReport, Long>, QuerydslPredicateExecutor<ProfileReport> {}
+public interface ProfileReportRepository extends JpaRepository<ProfileReport, Long> {}
